Update for compatibility with SQLA 1.4.43
- code updates to avoid "removed in 2.0" warnings
- local tests moved into their own files
- test_suite.py updates for current tests

Tested using fdb 2.0.2 against Firebird 3.0.10
and Firebird 4.0.2.
