+++ tox.ini
@@ -12,7 +12,7 @@ commands = flake8
 [flake8]
 # max-line-length = 88
 max-line-length = 79
-extend-ignore = E203,D400,D100,D101,D102,D104,I100,I101,I201,G002,G010
+extend-ignore = E203,F405,D400,D100,D101,D102,D104,I100,I101,I201,G002,G010
 exclude =
     .idea,
 	.tox,
